Make use of step command only when calling _privk3_XXX
Let a1 and a2 be two aspects, with a2 extending a1.
Let m1 be a method in a1 that is overriden in a2.

In a1, m1 is generated as a dispatch method that calls the
correct implementation using type checks.

With the previous implementation, having @Step on both m1
declarations (in a1 and in a2) would lead eventually to
executing two Step commands when calling m1 in a2. Yet this
doesn't make any sense since it is a single method call.

This commit solves this by only creating a step command when the
implementation of the method is called (ie. _privk3_XXX), and not
when executing the dispatch if statements of m1 in a1.